+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 🐞 Fixes
 
+- [#470](https://github.com/estruyf/vscode-front-matter/issues/470): Fix `initialize project` dashboard description
+
 ## [8.2.0] - 2022-12-08 - [Release notes](https://beta.frontmatter.codes/updates/v8.2.0) 
 
 ### ✨ New features